Register command for hover actions with Telescope
This is a tricky one: users probably want the default hover
functionality intact, so to trigger hover actions, it might be more
convenient to do it manually.

So in order to integrate with Telescope, it is better to preserver the
original command and simply reset it for Telescope, passing the
appropriate handler.