Rename commands
For some reason, I have a bunch of commands not related to rust-analyzer
(or even LSP at all) when editing Rust files (for example, `RustFmt`,
`RustPlay`, etc), and I don't think I use any plugins for Rust, so those
commands might just have come along Neovim itself.

So, in order to avoid confusion, and also to kind of follow the pattern
used in nvim-lspconfig, which is to start command names with the LSP
server's name, all commands now start with "RustAnalyzer", since this
plug-in is tightly coupled to rust-analyzer anyway.