fix(hover_actions): use current window to gather cursor position
This is to support scenarios where the buffer that was loaded into the
hover floating window has been moved into another window (for example a
preview window).

Instead of assuming that the buffer is still viewed in the floating
window, we now just use the current window to gather the cursor position.