feat(dap): std source map and custom dap configuration
Add a source map for libstd. Previously, stepping into the standard library
functions were not possible but now by setting an option in the
configuration, users can choose to set up a mapping that will be used by
the adapter when stepping into the standard library.
Allow custom configuration to be passed to the adapter.