Fix inlay hints crash in simrat39/rust-tools.nvim#300
According to LSP spec,
[InlayHint.label](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#inlayHint)
can be either a string or an array. Previously, `rust-analyzer` was
always returning strings. Most likely after
[this recent PR](https://github.com/rust-lang/rust-analyzer/pull/13699)
it started returning arrays.

Based on my testing, it seems only inlay hints with `kind` 1
are returning arrays but I am preemptively handling `kind` 2
as well in case it changes in the future.