fix(inlay_hints): disable strict_indexing for nvim_buf_get_lines
The last parameter of vim.api.nvim_buf_get_lines() is a boolean called
"strict_indexing". It causes the following behaviour:
- When strict_indexing is "true" out of bounds indices cause an error
- When strict_indexing is "false" out of bounds indices are clamped to
the nearest valid index.

Now nvim_buf_get_lines() returns an array, which might be empty. So
attempting to access the first element (our line) could evaluate to nil.

So we use "" as the fallback line string value.